+++ lbz/rest/api_gateway_event.py
@@ -11,20 +11,21 @@ class APIGatewayEvent(dict):
         self,
         resource_path: str,
         method: str,
-        body: dict = None,
-        query_params: dict = None,
-        path_params: dict = None,
-        headers: dict = None,
+        body: Optional[dict] = None,
+        query_params: Optional[dict] = None,
+        path_params: Optional[dict] = None,
+        headers: Optional[dict] = None,
     ) -> None:
         super().__init__()
 
         self._extract_query_params(query_params)
         self["resource"] = resource_path
+        self["httpMethod"] = method
         self["pathParameters"] = {} if path_params is None else path_params
         self["path"] = resource_path.format(**self.get("pathParameters", {}))
         self["method"] = method
         self["body"] = {} if body is None else body
-        self["headers"] = DEFAULT_HEADER if headers is None else headers
+        self["headers"] = DEFAULT_HEADERS if headers is None else headers
         self["queryStringParameters"] = query_params
         self["multiValueQueryStringParameters"] = query_params
         self["requestContext"] = {
@@ -36,9 +37,10 @@ class APIGatewayEvent(dict):
         self["stageVariables"] = {}
 
     def __repr__(self) -> str:
-        return f"<API Gateway Event {self['method']} @ {self['path']} body: {self['body']}>"
+        return f"<API Gateway Event {self['method']} @ {self['path']}>"
 
-    def _extract_query_params(self, query_params: Optional[dict]) -> None:
+    @staticmethod
+    def _extract_query_params(query_params: Optional[dict]) -> None:
         if query_params:
             for key, value in query_params.items():
                 if not isinstance(value, list):